Struct cbor::decoder::Config

pub struct Config {
    pub max_len_array: usize,
    pub max_len_bytes: usize,
    pub max_len_text: usize,
    pub max_size_map: usize,
    pub max_nesting: usize,
    pub skip_tags: bool,
    pub check_tags: bool,
}

Config contains various settings which limit resource consumption or enable certain validation options. Please note that the various maximum length/size values apply to an individual element only.

This is mainly to prevent attackers from providing CBOR values whose length is larger than the available memory. In combination the memory consumption can still become large and it is best to limit the incoming bytes to a specific upper bound, e.g. by using std::io::Take.

Fields

max_len_array

Maximum number of array elements

max_len_bytes

Maximum length of a byte string

max_len_text

Maximum length of a string

max_size_map

Maximum number of object fields

max_nesting

Maximum recursion steps when decoder Values

skip_tags

Ignore Tags when decoding Values

check_tags

Validate Value type matches Tag.

Methods

impl Config

fn default() -> Config

Create default configuration with

max_len_array = 1000

max_len_bytes = 5 MB

max_len_text = 5 MB

max_size_map = 1000

max_nesting = 16

skip_tags = false

check_tags = true