Struct cbor::decoder::GenericDecoder

pub struct GenericDecoder<R> {
    // some fields omitted
}

A generic decoder decodes arbitrary CBOR into a Value AST.

Methods

impl<R: ReadBytesExt> GenericDecoder<R>

fn new(c: Config, r: R) -> GenericDecoder<R>

fn from_decoder(d: Decoder<R>) -> GenericDecoder<R>

fn into_inner(self) -> Decoder<R>

fn borrow_mut(&mut self) -> &mut Decoder<R>

fn value(&mut self) -> DecodeResult<Value>

Decode into a Value, i.e. an intermediate representation which can be further deconstructed using a Cursor. This supports indefinite decoding as well as tag validation (if not disabled).