Enum cbor::Cbor

pub enum Cbor {
    Break,
    Undefined,
    Null,
    Bool(bool),
    Unsigned(CborUnsigned),
    Signed(CborSigned),
    Float(CborFloat),
    Bytes(CborBytes),
    Unicode(String),
    Array(Vec<Cbor>),
    Map(HashMap<String, Cbor>),
    Tag(CborTag),
}

CBOR abstract syntax.

This type can represent any data item described in the CBOR specification with some restrictions. Namely, CBOR maps are limited to Unicode string keys.

Note that this representation distinguishes the size of an encoded number.

Variants

Break

A code used to signify the end of an indefinite length data item.

Undefined

An undefined data item (major type 7, value 23).

Null

A null data item (major type 7, value 22).

Bool

A boolean data item (major type 7, values 20 or 21).

Unsigned

An unsigned integer (major type 0).

Signed

A negative integer (major type 1).

Float

An IEEE 754 floating point number (major type 7).

Bytes

A byte string (major type 2).

Unicode

A Unicode string (major type 3).

Array

An array (major type 4).

Map

A map (major type 5).

Tag

A tag (major type 6).

Methods

impl Cbor

fn decode<D: Decodable>(self) -> CborResult<D>

Decode a single CBOR value.

fn tag(&self) -> Option<u64>

If this is a CBOR tag, return the tag number.