Enum cbor::CborSigned

pub enum CborSigned {
    Int8(i8),
    Int16(i16),
    Int32(i32),
    Int64(i64),
}

A negative integer (major type 1).

Note that while the number -255 can be encoded using two bytes as a CBOR uint8, it is outside the range of numbers allowed in i8. Therefore, when CBOR data is decoded, -225 is stored as in i16 in memory.

Variants

Int8

Negative 8 bit integer.

Int16

Negative 16 bit integer.

Int32

Negative 32 bit integer.

Int64

Negative 64 bit integer.

Methods

impl CborSigned

fn into_i64(self) -> i64

Return the underlying value as a i64.