Enum cbor::ReadError

pub enum ReadError {
    TypeMismatch {
        expected: Type,
        got: Type,
    },
    InvalidAddValue {
        ty: Type,
        val: u8,
    },
    Unassigned {
        major: u8,
        add: u8,
    },
    Reserved {
        major: u8,
        add: u8,
    },
    Other(String),
}

An error produced by reading CBOR data.

Variants

TypeMismatch

An error for when the expected type does not match the received type.

Fields

expected

Expected CBOR type.

got

Received CBOR type.

InvalidAddValue

When the additional information is inconsistent with the major type.

Fields

ty

CBOR type.

val

Additional information value.

Unassigned

The value found is unassigned.

Fields

major

CBOR major type value.

add

Additional information value.

Reserved

The value found is reserved.

Fields

major

CBOR major type value.

add

Additional information value.

Other

Some other error occurred.