Struct cbor::CborTag

pub struct CborTag {
    pub tag: u64,
    pub data: Box<Cbor>,
}

A tag (major type 6).

Note that if you want to encode a tag, you should use the CborTagEncode type and not this type. This type is only useful when your manually expecting the structure of a CBOR data item.

Fields

tag

The tag number.

You can see a list of currently assigned tag numbers here: http://www.iana.org/assignments/cbor-tags/cbor-tags.xhtml

Note that it is OK to choose your own tag number for your own application specific purpose, but it should probably be one that is currently unassigned in the IANA registry.

data

The data item, represented in terms of CBOR abstract syntax.