Struct cbor::value::Cursor

pub struct Cursor<'r> {
    // some fields omitted
}

A Cursor allows conventient navigation in a Value AST. Values can be converted to native Rust types if possible and collections can be traversed using at or get.

Methods

impl<'r> Cursor<'r>

fn new(v: &'r Value) -> Cursor<'r>

fn at(&self, i: usize) -> Cursor<'r>

fn get(&self, k: Key) -> Cursor<'r>

fn field(&self, s: &str) -> Cursor<'r>

fn value(&self) -> Option<&Value>

fn opt(&self) -> Option<Cursor<'r>>

fn maybe(&self) -> Option<Cursor<'r>>

fn bool(&self) -> Option<bool>

fn bytes(&self) -> Option<&Bytes>

fn bytes_plain(&self) -> Option<&Vec<u8>>

fn bytes_chunked(&self) -> Option<&LinkedList<Vec<u8>>>

fn text(&self) -> Option<&Text>

fn text_plain(&self) -> Option<&String>

fn text_chunked(&self) -> Option<&LinkedList<String>>

fn float32(&self) -> Option<f32>

fn float64(&self) -> Option<f64>

fn u8(&self) -> Option<u8>

fn u16(&self) -> Option<u16>

fn u32(&self) -> Option<u32>

fn u64(&self) -> Option<u64>

fn i8(&self) -> Option<i8>

fn i16(&self) -> Option<i16>

fn i32(&self) -> Option<i32>

fn i64(&self) -> Option<i64>