Crate cbor

CBOR (RFC 7049) encoder and decoder implementations.

Re-exports

pub use decoder::Config;

pub use decoder::Decoder;

pub use decoder::DecodeError;

pub use decoder::DecodeResult;

pub use decoder::GenericDecoder;

pub use decoder::opt;

pub use decoder::maybe;

pub use decoder::or_break;

pub use encoder::Encoder;

pub use encoder::EncodeError;

pub use encoder::EncodeResult;

pub use encoder::GenericEncoder;

Modules

decoder
CBOR (RFC 7049) decoder implementation supporting the following features:

encoder
CBOR (RFC 7049) encoder implementation.

skip
Skip trait to allow efficient skipping of consecutive bytes.

slice
ReadSlice trait to allow efficient reading of slices without copying.

types
CBOR types and tags definitions.

value
This module defines the generic Value AST as well as several other types to represent CBOR values. A Cursor can be used to deconstruct and traverse a Value.