Crate cbor

CBOR (RFC 7049) encoder and decoder implementations.

Reexports

pub use decoder::{Config, Decoder, DecodeError, DecodeResult, GenericDecoder};

pub use decoder::{opt, maybe, or_break};

pub use encoder::{Encoder, EncodeError, EncodeResult, GenericEncoder};

Modules

decoder
CBOR (RFC 7049) decoder implementation supporting the following features:

encoder
CBOR (RFC 7049) encoder implementation.

random
Generate random cbor::Values.

types
CBOR types and tags definitions.

value
This module defines the generic Value AST as well as several other types to represent CBOR values. A Cursor can be used to deconstruct and traverse a Value.