Enum cbor::CborError

pub enum CborError {
    Io(Error),
    Decode(ReadError),
    Encode(WriteError),
    AtOffset {
        kind: ReadError,
        offset: usize,
    },
    UnexpectedEOF,
}

Errors that can be produced by a CBOR operation.

Variants

Io

An error as a result of an underlying IO operation.

Decode

An error from the type based decoder.

Encode

An error from the type based encoder.

AtOffset

An error reading CBOR at a particular offset.

For example, if the data in "additional information" is inconsistent with the major type.

Fields

kind

The exact read error.

offset

The byte offset at which the error occurred.

UnexpectedEOF

EOF is found but more bytes were expected to decode the next data item.

EOF is triggered when the underlying reader returns 0 bytes.